[moved number formatter to utils, reused the same across different components][added extra props to denote price decimals][references issue #8]
