+++ lisk-dex-electron/src/OrderbookEntry.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import './OrderbookEntry.css';
+import formatThousands from './Utils'
 
 export default class OrderbookEntry extends React.Component {
   constructor(props) {
@@ -17,11 +18,18 @@ export default class OrderbookEntry extends React.Component {
       return `linear-gradient(to right, #286113 ${percentOfMaxSize}%, rgba(0,0,0,0) 1%)`;
     }
   }
+  
 
   render() {
+
+    const formattedPrice = this.props.price.toFixed(this.props.priceDecimals)
+    const size = (this.props.size / this.props.whole).toFixed(this.props.sizeDecimals)
+    const formattedSize = formatThousands(size, ',')
+
     return (
+
       <div style={{ background: this.bgCSS(), marginTop: '1px', marginBottom: '1px' }} className="orderLine">
-        {this.props.price}, {(this.props.size / this.props.whole).toFixed(this.props.decimals)}
+        {formattedPrice} &nbsp;|&nbsp; {formattedSize}
       </div>
     );
   }
