+++ lisk-dex-electron/src/PriceHistoryChart.js
@@ -2,34 +2,10 @@ import React from 'react';
 import {
   ComposedChart, Bar, Line, XAxis, YAxis, CartesianGrid, Tooltip
 } from 'recharts';
+import formatThousands from './Utils'
 
 const DEFAULT_VOLUME_DISPLAY_HEIGHT_RATIO = .2;
 
-const formatThousands = function (num, separator) {
-	separator = separator || ',';
-	let numParts = [];
-	let paddingZero = '0';
-
-	let fractionDecimals = num.toString().split('.')[1] || '';
-	if (fractionDecimals.length) {
-		fractionDecimals = '.' + fractionDecimals;
-	}
-	let remaining = Math.floor(num);
-	if (remaining === 0) {
-		return remaining + fractionDecimals;
-	}
-
-	let lastDigits;
-	while (remaining !== 0) {
-		lastDigits = (remaining % 1000).toString();
-		remaining = Math.floor(remaining / 1000);
-		if (remaining !== 0) {
-			lastDigits =  lastDigits + paddingZero.repeat(3 - lastDigits.length);
-		}
-		numParts.push(lastDigits);
-	}
-	return numParts.reverse().join(separator) + fractionDecimals;
-}
 
 const PriceTooltip = ({ active, payload, label }) => {
 	if (!active || !payload) {
